% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide.R
\name{tide}
\alias{tide}
\title{Edit Manually A Dataframe And Generate Reproducible Code}
\usage{
tide(df, clip = TRUE)
}
\arguments{
\item{df}{A data.frame to be edited.}

\item{clip}{Logical, defaults to TRUE. Copy generated code to the clipboard?}
}
\value{
A data.frame. Optionally adds a string to the clipboard.
}
\description{
Calls the \code{\link{edit}} function to invoke a spreadsheet-like
data-editor window where you can edit a data.frame object by hand. Returns
the edited object, but also generates code that can can recreate the amended
object from the original.
}
\details{
As noted in \code{?dataentry}, 'the data entry editor is only
available on some platforms and GUIs' and 'The details of interface to
the data grid may differ by platform and GUI'.
}
\examples{
\dontrun{tide(head(beaver1))}
}
