% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide-package.R
\docType{package}
\name{tide-package}
\alias{tide-package}
\title{tide: Manually Edit A Dataframe, Get Code to Reproduce It}
\description{
Adjust a data.frame manually with R's inbuilt spreadsheet-like data editor and have simple base-R code returned that reproduces the changes.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
