% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide.R
\name{tide}
\alias{tide}
\title{Edit Manually A Dataframe And Generate Reproducible Code}
\usage{
tide(dat, copy_to_clipboard = TRUE)
}
\arguments{
\item{dat}{A data.frame to be edited. Must be the name of the data.frame
object only.}

\item{copy_to_clipboard}{Logical, defaults to TRUE. Copy generated code to
the clipboard?}
}
\value{
A data.frame. Optionally adds a string to the clipboard.
}
\description{
Calls the \link[utils]{edit} function to invoke a spreadsheet-like
data-editor window where you can edit a data.frame object 'by hand'. Returns
the edited object and optionally adds code to the clipboard that can recreate
the edited object from the original.
}
\details{
As noted in \link[utils]{dataentry}, 'the data entry editor is only
available on some platforms and GUIs' and 'The details of interface to
the data grid may differ by platform and GUI'.
}
\examples{
\dontrun{beaver1_edited <- tide(beaver1)}
}
